wip: cheat

+++ install-cheat.sh
@@ -80,12 +80,14 @@ fi
 # (tar.gz)
 cd /usr/local/bin
 
+if [ ! -f /usr/local/bin/cheat ]; then
+  curl -L "https://github.com/cheat/cheat/releases/download/${CHEAT_VERSION}/cheat-${SYSTEM}-${HARDWARE}.gz" > cheat-${SYSTEM}-${HARDWARE}.gz
+  gzip -d cheat-${SYSTEM}-${HARDWARE}.gz
+  mv cheat-${SYSTEM}-${HARDWARE} cheat
+  chmod +x /usr/local/bin/cheat
+fi
 
-curl -L "https://github.com/cheat/cheat/releases/download/${CHEAT_VERSION}/cheat-${SYSTEM}-${HARDWARE}.gz" > cheat-${SYSTEM}-${HARDWARE}.gz
-gzip -d cheat-${SYSTEM}-${HARDWARE}.gz
-mv cheat-${SYSTEM}-${HARDWARE} cheat
-# https://github.com/cheat/cheat/releases/download/4.1.0/cheat-linux-amd64.gz
-chmod +x /usr/local/bin/cheat
+[ ! -f ~/.config/cheat ] && mkdir -p ~/.config/cheat; cheat --init > ~/.config/cheat/conf.yml
 
 cd -
